Improved data-merging for widgets
Before, merging of data from the HTTP request or the object model was done without the possibility to ignore either source (or to provide an extra source).

This commit introduces new methods for configuring how and from where a widget’s data comes from. The ‘TableWidget’ and ‘ObjectFormWidget’ have been updated to take advantage of these configurable/optional extra data sources.

Basic customization of additional data sources:

```
{
  "type": "charcoal/admin/widget/table",
  "collection_ident": “foobar”,
  "obj_type": “charcoal/cms/section",
  "data_sources": [ "request", "object" ]
}
```

Where `request` and `object` are valid sources (defined in `AdminWidget::acceptedDataSources()` which are invoked as `dataFrom<Source>()` on the widgets, if present.

A “merge strategy” can be defined by passing the data sources as an associative array of callables.

```
"data_sources": { "request": "myCustomDataSource", "object": "array_merge_recursive" }
```

Where `myCustomDataSource` and `array_merge_recursive` are functions used to merge data. The `myCustomDataSource`, in the context of `TableWidget` and `ObjectFormWidget`, is a method you'd define on your `Section` model.