Added support for hidden property controls
Useful for submitting data to an object model that isn’t necessarily open to user input and a disabled or read-only control would be out of place.

Hidden properties can be assigned to the form structure or form group structure with “properties” or “hidden_properties” to be good and proper.

Changes:
- Added setter/getter for ‘FormWidget::$hiddenProperties’;
- Modified ‘FormWidget::addFormProperty()’ to reassign hidden controls;
- Modified templates to output any hidden controls from the model;
- Modified ‘FormPropertyWidget::inputType()’ to detect if the property is hidden;
- Added ‘FormPropertyWidget::hidden()’ to check if the input type is hidden or if the property is hidden;
- Modified ‘FormGroupWidget’ and ‘StructureFormGroup’ to reassign hidden controls to the form;