Selectize Update :
Add support for a create button using the option allow_create

This add-on is useful when the label cannot be directly mapped to a property.
Pressing a button to create an object on the fly will ease the experience in those scenarios, no need to enter text.
You can also disable the option to create item using the usual method to make sure to avoid confusion.

```
"allow_create": true,
"selectize_options": {
   "create": false
}
```