Tinymce changes :
	- Add default config for automatic_upload and image_upload_url
	- Add Tinymce/UploadImageAction To create a new file from tinymce directly.

This allow a default setup to create a server image of uploaded files by the mean of drag n drop functionality instead of saving images directly into database.
Fixes bug when uploading size heavy images using the said method which would results in unclosed html img tag due to the fact that the image exceeds the memory limit of the database table column.