Add support for escaping values on properties
Added support to escape display and input values. The feature is disabled (false), by default. If the feature is enabled (true), the default function is `htmlspecialchars()`. The escape option is a function name or expression and optional parameters.

Added:
- Option "display_escape" to escape display properties.
- Option "input_escape" to escape input properties.
- Escaping of message in Message display property (HTML is expected).
- Escaping to Read-Only input property (HTML is expected).

Usage:

Useful for escaping special characters such as "<" and ">" in email addresses ("<charcoal@locomotive.ca>").

```json
"email_address": {
    "type": "email",
    "input_type": "charcoal/admin/property/input/readonly",
    "display_escape": true
}
```

```php
$input->setEscape('htmlentities');
$input->setEscape(function ($value) {
    return htmlentities($value, ENT_QUOTES, 'UTF-8');
});
$input->setEscape([
    'function'   => 'htmlentities',
    'parameters' => [ ENT_QUOTES | ENT_IGNORE, 'UTF-8' ],
]);
```